backend/pyproject.toml
[tool.poetry]
name = "psi-backend"
version = "1.0.0"
description = "Psi - Emotion-based Wellness Platform Backend"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
pymongo = "^4.6.1"
redis = "^5.0.1"
motor = "^3.3.2"
anthropic = "^0.18.0"
ultralytics = "^8.1.0"
opencv-python = "^4.9.0"
numpy = "^1.26.0"
scikit-learn = "^1.4.0"
pillow = "^10.2.0"
boto3 = "^1.34.0"
python-dotenv = "^1.0.0"
alembic = "^1.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.26.0"
black = "^24.1.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
